api/requirements.txt
aiohttp==3.9.4
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
bleach==6.1.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
dnspython==2.6.1
email_validator==2.1.1
fastapi>=0.115.12
another-fastapi-jwt-auth==0.1.6
filelock==3.13.3
flake8==7.0.0
frozenlist==1.4.1
fsspec==2024.3.1
h11==0.14.0
httpcore==1.0.5
httpx==0.26.0
huggingface-hub==0.22.2
idna==3.7
importlib_metadata==7.1.0
iniconfig==2.0.0
Jinja2==3.1.3
MarkupSafe==2.1.5
mccabe==0.7.0
motor==3.3.2
mpmath==1.3.0
multidict==6.0.5
networkx==3.2.1
numpy==1.26.4
openai>=1.27.0
packaging<24.0
passlib==1.7.4
pluggy==1.4.0
pycodestyle==2.11.1
pydantic>=2.0.0,<3.0.0
pyflakes==3.2.0
PyJWT>=2.0.0
pymongo==4.7.2
pytest==8.0.0
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.32.2
safetensors==0.4.2
sniffio==1.3.1
starlette>=0.36.3
sympy==1.12
tiktoken==0.6.0
tokenizers==0.15.2
tqdm==4.66.4
transformers==4.39.3
python-multipart==0.0.9
python-socketio==5.11.2


typing_extensions==4.10.0
urllib3<2
uvicorn==0.25.0
gunicorn~=22.0.0
yarl==1.9.4
zipp==3.18.1
apscheduler==3.10.4
opensearch-py==2.5.0
nltk==3.8.1
langchain-community==0.2.6
langchain==0.2.6
sentence-transformers==2.7.0
unstructured==0.14.0
pypdf==4.2.0
python-docx==1.1.2


pip~=24.0
colorama~=0.4.6
setuptools~=70.3.0
six~=1.16.0
webencodings~=0.5.1
wsproto~=1.2.0
python-dateutil~=2.9.0.post0
greenlet~=3.0.3
SQLAlchemy~=1.4.52
neo4j~=5.22.0
langchain-experimental==0.0.62
jsonpath-python==1.0.6